smyrna: make 'bd{x|y}{Left|Right|Top|Bottom}' 'double's
This is part of a series of changes to standardize on `double` arithmetic in
Smyrna. This removes visual noise and improves accuracy. The computational cost
of `double` vs `float` arithmetic is no longer a significant factor in
contemporary environments.

+++ cmd/smyrna/smyrnadefs.h
@@ -281,8 +281,8 @@ enum {SCHEMACOUNT = 4}; // number of colors
 	int bdVisible;		//if borders are visible (boundries of the drawing,
 	/*border coordinates, needs to be calculated for each graph */
 
-	float bdxLeft, bdyTop;
-	float bdxRight, bdyBottom;
+	double bdxLeft, bdyTop;
+	double bdxRight, bdyBottom;
 
 	/*screen window size in 2d */
 	int w, h;
